Refactor deck-specific value names.
This deprecates and replaces the constants used for deck-specific
values with dynamically generated ones. This is actually closer to
how the original Engine software treats these names.

On the next major version we would actually remove the old
constants to clean up the code, but for now we can maintain the
pre-existing values. Newly added values will NOT be added as
constants to keep technical debt at a minimum.

The references in the example command have been replaced
respectively.