Make beacon work on windows.
We have to use a longer path to find the correct IP to report gRPC
on: Go through all interface nets and find the one where the
source can be expected to sit in.

This may actually be closer to the original software's behavior
since we only get one reply with the most fitting IP back instead
of multiple responses per interface reached.